requirements.txt
numpy
onnx
onnxruntime
tqdm
pyyaml
pydantic
strictyaml
loguru
pandas
opencv-python
matplotlib
seaborn